requirements.txt
requests
python-dotenv
beautifulsoup4
lxml
pytelegrambotapi
vinted-api-wrapper
